@@ -10,17 +10,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~17.3.12",
+    "@angular/animations": "~18.2.13",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "~17.3.12",
-    "@angular/compiler": "~17.3.12",
-    "@angular/core": "~17.3.12",
-    "@angular/forms": "~17.3.12",
-    "@angular/localize": "^17.3.12",
+    "@angular/common": "~18.2.13",
+    "@angular/compiler": "~18.2.13",
+    "@angular/core": "~18.2.13",
+    "@angular/forms": "~18.2.13",
+    "@angular/localize": "^18.2.13",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "~17.3.12",
-    "@angular/platform-browser-dynamic": "~17.3.12",
-    "@angular/router": "~17.3.12",
+    "@angular/platform-browser": "~18.2.13",
+    "@angular/platform-browser-dynamic": "~18.2.13",
+    "@angular/router": "~18.2.13",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "@popperjs/core": "^2.11.8",
     "angular-google-charts": "^16.0.2",
@@ -41,15 +41,15 @@
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.17",
+    "@angular-devkit/build-angular": "^18.2.19",
     "@angular-eslint/builder": "^20.0.0",
     "@angular-eslint/eslint-plugin": "^20.0.0",
     "@angular-eslint/eslint-plugin-template": "^20.0.0",
     "@angular-eslint/schematics": "^20.0.0",
     "@angular-eslint/template-parser": "^20.0.0",
-    "@angular/cli": "^17.3.17",
-    "@angular/compiler-cli": "~17.3.12",
-    "@angular/language-service": "~17.3.12",
+    "@angular/cli": "^18.2.19",
+    "@angular/compiler-cli": "~18.2.13",
+    "@angular/language-service": "~18.2.13",
     "@types/datatables.net": "^1.10.18",
     "@types/datatables.net-select": "^1.2.6",
     "@types/jasmine": "~3.6.0",
@@ -66,7 +66,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^18.2.1",
     "ts-node": "~7.0.0",
     "typescript": "^5.4.5"
   },
