@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { APP_INITIALIZER, CUSTOM_ELEMENTS_SCHEMA, ErrorHandler, NgModule } from '@angular/core';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
@@ -105,93 +105,87 @@ const appRoutes: Routes = [
   { path: '**', component: HomeComponent, canActivate: [AuthGuard] },
 ];
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    AppHeaderComponent,
-    HomeComponent,
-    GenGeneTableComponent,
-    GenGeneTablePanelComponent,
-    GeneTableSelectorComponent,
-    DropdownDirective,
-    SeqModalComponent,
-    GeneBrowserComponent,
-    GeneBrowserPanelComponent,
-    GeneTruncPipe,
-    RepSampleComponent,
-    RepSamplePanelComponent,
-    ColumnSorterComponent,
-    RepSampleInfoComponent,
-    RepGeneTableComponent,
-    RepGeneTablePanelComponent,
-    GenSampleComponent,
-    GenSamplePanelComponent,
-    GenSampleInfoComponent,
-    ReportsComponent,
-    ReportParamsDialogComponent,
-    RepGeneNotesComponent,
-    SafePipe,
-    ReportErrorDialogComponent,
-    QuickRefComponent,
-    UserGuideComponent,
-    RepExploreDataComponent,
-    RepExploreDataPanelComponent,
-    LicensingComponent,
-    ReportRunDialogComponent,
-    AuthComponent,
-    GenExploreDataComponent,
-    GenExploreDataPanelComponent,
-    RefbookComponent,
-    RefbookPanelComponent,
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    HttpClientModule,
-    FormsModule,
-    ReactiveFormsModule,
-    NgbModule,
-    MatInputModule,
-    MatTableModule,
-    MatPaginatorModule,
-    MatSortModule,
-    MatProgressSpinnerModule,
-    RouterModule.forRoot(appRoutes),
-    ApiModule.forRoot(apiConfigFactory),
-    NgMultiSelectDropDownModule.forRoot(),
-    FilterModule,
-    MatIconModule,
-    MatMenuModule,
-    MatCheckboxModule,
-    DragDropModule,
-    MatCardModule,
-    ObserversModule,
-    ResizableModule,
-    GoogleChartsModule,
-    MatTooltipModule,
-  ],
-  providers: [
-    RequestCache,
-    AuthService,
-    ReportsService,
-    GenomicService,
-    RepseqService,
-    SystemService,
-    ReportRunService,
-    { provide: BASE_PATH, useValue: environment.apiBasePath },
-    { provide: HTTP_INTERCEPTORS, useClass: CachingInterceptor, multi: true },
-    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptorService, multi: true },
-    { provide: RouteReuseStrategy, useClass: CustomReuseStrategy },
-    { provide: APP_INITIALIZER, useFactory: appInitializer, deps: [AuthService], multi: true, },
-  ],
-  bootstrap: [AppComponent],
-  schemas: [
-    // Add CUSTOM_ELEMENTS_SCHEMA to suppress errors about unknown elements in templates
-    // This is needed for the ResizableModule directives to work with material components
-    // We would need to refactor the code to use the proper directives, but this is a workaround for now
-    // during the Angular upgrade
-    CUSTOM_ELEMENTS_SCHEMA
-  ],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        AppHeaderComponent,
+        HomeComponent,
+        GenGeneTableComponent,
+        GenGeneTablePanelComponent,
+        GeneTableSelectorComponent,
+        DropdownDirective,
+        SeqModalComponent,
+        GeneBrowserComponent,
+        GeneBrowserPanelComponent,
+        GeneTruncPipe,
+        RepSampleComponent,
+        RepSamplePanelComponent,
+        ColumnSorterComponent,
+        RepSampleInfoComponent,
+        RepGeneTableComponent,
+        RepGeneTablePanelComponent,
+        GenSampleComponent,
+        GenSamplePanelComponent,
+        GenSampleInfoComponent,
+        ReportsComponent,
+        ReportParamsDialogComponent,
+        RepGeneNotesComponent,
+        SafePipe,
+        ReportErrorDialogComponent,
+        QuickRefComponent,
+        UserGuideComponent,
+        RepExploreDataComponent,
+        RepExploreDataPanelComponent,
+        LicensingComponent,
+        ReportRunDialogComponent,
+        AuthComponent,
+        GenExploreDataComponent,
+        GenExploreDataPanelComponent,
+        RefbookComponent,
+        RefbookPanelComponent,
+    ],
+    bootstrap: [AppComponent],
+    schemas: [
+        // Add CUSTOM_ELEMENTS_SCHEMA to suppress errors about unknown elements in templates
+        // This is needed for the ResizableModule directives to work with material components
+        // We would need to refactor the code to use the proper directives, but this is a workaround for now
+        // during the Angular upgrade
+        CUSTOM_ELEMENTS_SCHEMA
+    ], imports: [BrowserModule,
+        BrowserAnimationsModule,
+        FormsModule,
+        ReactiveFormsModule,
+        NgbModule,
+        MatInputModule,
+        MatTableModule,
+        MatPaginatorModule,
+        MatSortModule,
+        MatProgressSpinnerModule,
+        RouterModule.forRoot(appRoutes),
+        ApiModule.forRoot(apiConfigFactory),
+        NgMultiSelectDropDownModule.forRoot(),
+        FilterModule,
+        MatIconModule,
+        MatMenuModule,
+        MatCheckboxModule,
+        DragDropModule,
+        MatCardModule,
+        ObserversModule,
+        ResizableModule,
+        GoogleChartsModule,
+        MatTooltipModule], providers: [
+        RequestCache,
+        AuthService,
+        ReportsService,
+        GenomicService,
+        RepseqService,
+        SystemService,
+        ReportRunService,
+        { provide: BASE_PATH, useValue: environment.apiBasePath },
+        { provide: HTTP_INTERCEPTORS, useClass: CachingInterceptor, multi: true },
+        { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptorService, multi: true },
+        { provide: RouteReuseStrategy, useClass: CustomReuseStrategy },
+        { provide: APP_INITIALIZER, useFactory: appInitializer, deps: [AuthService], multi: true, },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule {
 }
