@@ -1,7 +1,7 @@
 // Collect recent posts from Wordpress
 
 import {Injectable, OnInit} from '@angular/core';
-import {HttpClient} from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import {BehaviorSubject, EMPTY, Observable, Subject, throwError} from 'rxjs';
 import {map, catchError} from "rxjs/operators";
 import {retryWithBackoff} from "../shared/retry_with_backoff";
